Refactor page generation code into page package
