+++ cli/cli.go
@@ -3,96 +3,27 @@
 // Use of this source code is governed by the MIT
 // license that can be found in the LICENSE.md file.
 
-// TODO: Refactor this package.
-
-// Package cli implements command line interface of gen.
+// Package cli implements command line interface.
 package cli // import "astrophena.me/gen/cli"
 
 import (
 	"fmt"
-	"html/template"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"astrophena.me/gen/buildinfo"
 	"astrophena.me/gen/fileutil"
+	"astrophena.me/gen/page"
 	"astrophena.me/gen/scaffold"
 
-	"github.com/oxtoacart/bpool"
-	"github.com/tdewolff/minify/v2"
-	"github.com/tdewolff/minify/v2/css"
-	"github.com/tdewolff/minify/v2/html"
 	"github.com/urfave/cli/v2"
 )
 
-const bufpoolSize = 48
+var minCSS string
 
-var (
-	bufpool *bpool.BufferPool
-	m       *minify.M
-	minCSS  string
-	tpl     *template.Template
-)
-
-type page struct {
-	URI         string
-	Title       string
-	Description string
-	Body        string
-	MetaTags    map[string]string
-	CSS         string
-
-	template string
-	filename string
-}
-
-func (p *page) Generate(dst string) (err error) {
-	dir := filepath.Join(dst, filepath.Dir(p.URI))
-	if err := fileutil.Mkdir(dir); err != nil {
-		return err
-	}
-
-	var (
-		tbuf = bufpool.Get()
-		mbuf = bufpool.Get()
-	)
-	defer bufpool.Put(tbuf)
-	defer bufpool.Put(mbuf)
-
-	if err := tpl.ExecuteTemplate(tbuf, p.template, p); err != nil {
-		return err
-	}
-
-	if err := m.Minify("text/html", mbuf, tbuf); err != nil {
-		return err
-	}
-
-	f, err := os.Create(filepath.Join(dst, p.URI))
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	if _, err := mbuf.WriteTo(f); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func init() {
-	m = minify.New()
-	m.AddFunc("text/html", html.Minify)
-	m.AddFunc("text/css", css.Minify)
-
-	bufpool = bpool.NewBufferPool(bufpoolSize)
-}
-
-// Run invokes command line interface of gen.
+// Run invokes command line interface.
 func Run() {
 	app := &cli.App{
 		Name:    "gen",
@@ -114,16 +45,16 @@ func Run() {
 		},
 		Commands: []*cli.Command{
 			{
-				Name:    "new",
-				Aliases: []string{"n"},
-				Usage:   "Creates a new site in the provided directory",
-				Action:  newCmd,
-			},
-			{
 				Name:    "build",
 				Aliases: []string{"b"},
 				Usage:   "Performs a one off site build",
 				Action:  build,
+			},
+			{
+				Name:    "create",
+				Aliases: []string{"c"},
+				Usage:   "Creates a new site in the provided directory",
+				Action:  create,
 			},
 			{
 				Name:    "serve",
@@ -140,10 +71,10 @@ func Run() {
 				Action: serve,
 			},
 			{
-				Name:    "clean",
-				Aliases: []string{"c"},
+				Name:    "remove",
+				Aliases: []string{"r"},
 				Usage:   "Removes all generated files",
-				Action:  clean,
+				Action:  remove,
 			},
 		},
 	}
@@ -152,20 +83,6 @@ func Run() {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-}
-
-func newCmd(c *cli.Context) (err error) {
-	dst := c.Args().Get(0)
-
-	if dst == "" {
-		return fmt.Errorf("directory is required, but not provided")
-	}
-
-	if err := scaffold.Create(dst); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func build(c *cli.Context) (err error) {
@@ -181,43 +98,20 @@ func build(c *cli.Context) (err error) {
 		css = filepath.Join(assetsDir, "sitewide.css")
 
 		start = time.Now()
-
-		tplFuncs = template.FuncMap{
-			"css": func(s string) template.CSS {
-				return template.CSS(s)
-			},
-			"html": func(s string) template.HTML {
-				return template.HTML(s)
-			},
-			"year": func() int {
-				return time.Now().Year()
-			},
-			"version": buildinfo.TplFunc(),
-		}
 	)
 
-	tpl = template.New("").Funcs(tplFuncs)
+	fmt.Printf("Building into %s.\n", dst)
+
+	if err := remove(c); err != nil {
+		return err
+	}
 
 	tpls, err := fileutil.Files(templatesDir, "html")
 	if err != nil {
 		return err
 	}
 
-	for _, t := range tpls {
-		f, err := ioutil.ReadFile(t)
-		if err != nil {
-			return err
-		}
-
-		tpl, err = tpl.Parse(string(f))
-		if err != nil {
-			return err
-		}
-	}
-
-	fmt.Printf("Building into %s.\n", dst)
-
-	if err := clean(c); err != nil {
+	if err := page.ParseTemplates(tpls); err != nil {
 		return err
 	}
 
@@ -235,14 +129,14 @@ func build(c *cli.Context) (err error) {
 	}
 
 	if fileutil.Exists(css) {
-		minCSS, err = minifyCSS(css)
+		minCSS, err = page.Minify("text/css", css)
 		if err != nil {
 			return err
 		}
 	}
 
 	for _, f := range content {
-		p, err := parseFile(f)
+		p, err := page.ParseFile(f, minCSS)
 		if err != nil {
 			return err
 		}
@@ -255,6 +149,20 @@ func build(c *cli.Context) (err error) {
 	}
 
 	fmt.Printf("Built in %v.\n", time.Since(start))
+
+	return nil
+}
+
+func create(c *cli.Context) (err error) {
+	dst := c.Args().Get(0)
+
+	if dst == "" {
+		return fmt.Errorf("directory is required, but not provided")
+	}
+
+	if err := scaffold.Create(dst); err != nil {
+		return err
+	}
 
 	return nil
 }
@@ -277,7 +185,7 @@ func serve(c *cli.Context) (err error) {
 		return err
 	}
 
-	fmt.Printf("Serving site on port %v.\n", port)
+	fmt.Printf("Serving on port %v.\n", port)
 	if err := srv.ListenAndServe(); err != nil {
 		if err != http.ErrServerClosed {
 			return err
@@ -287,7 +195,7 @@ func serve(c *cli.Context) (err error) {
 	return nil
 }
 
-func clean(c *cli.Context) (err error) {
+func remove(c *cli.Context) (err error) {
 	dst := c.String("destination")
 
 	if fileutil.Exists(dst) {
@@ -297,69 +205,4 @@ func clean(c *cli.Context) (err error) {
 	}
 
 	return nil
-}
-
-func parseFile(filename string) (*page, error) {
-	b, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-	content := string(b)
-	separator := "\n---\n"
-	position := strings.Index(content, separator)
-	if position <= 0 {
-		return nil, fmt.Errorf("%s: no header section detected", filename)
-	}
-
-	header := content[:position]
-	p := &page{
-		Body:     content[position+len(separator):],
-		filename: filename,
-		MetaTags: make(map[string]string),
-		CSS:      minCSS,
-	}
-
-	for _, line := range strings.Split(header, "\n") {
-		switch {
-		case strings.HasPrefix(line, "title: "):
-			p.Title = line[7:]
-		case strings.HasPrefix(line, "description: "):
-			p.MetaTags["description"] = line[13:]
-		case strings.HasPrefix(line, "template: "):
-			p.template = line[10:]
-		case strings.HasPrefix(line, "uri: "):
-			p.URI = line[5:]
-		case strings.HasPrefix(line, "meta-tag: "):
-			t := strings.Split(line[10:], "=")
-			p.MetaTags[t[0]] = t[1]
-		}
-	}
-
-	if p.Title == "" || p.template == "" || p.URI == "" {
-		return nil, fmt.Errorf("%s: missing required header parameter (title, template, uri)", filename)
-	}
-
-	if tpl.Lookup(p.template) == nil {
-		return nil, fmt.Errorf("%s: the template %s specified is not defined", filename, p.template)
-	}
-
-	return p, nil
-}
-
-func minifyCSS(path string) (css string, err error) {
-	if !fileutil.Exists(path) {
-		return "", fmt.Errorf("%s: no such file", path)
-	}
-
-	b, err := ioutil.ReadFile(path)
-	if err != nil {
-		return "", err
-	}
-
-	s, err := m.Bytes("text/css", b)
-	if err != nil {
-		return "", fmt.Errorf("%s: failed to minify: %w", path, err)
-	}
-
-	return string(s), nil
 }
