+++ go.mod
@@ -3,7 +3,6 @@ module astrophena.me/gen
 go 1.14
 
 require (
-	github.com/oxtoacart/bpool v0.0.0-20190530202638-03653db5a59c
 	github.com/tdewolff/minify/v2 v2.7.2
 	github.com/urfave/cli/v2 v2.1.1
 )
